requirements.txt
attrs==16.3.0
beautifulsoup4==4.5.3
bs4==0.0.1
cffi==1.9.1
chardet==3.0.2
constantly==15.1.0
cryptography==1.7.1
cssselect==1.0.0
defusedxml==0.4.1
Django==1.10.2
django-bootstrap-form==3.2.1
django-contrib-comments==1.7.3
feedparser==5.2.1
idna==2.2
incremental==16.10.1
Janome==0.2.8
lxml==3.6.4
numpy==1.11.2
oauthlib==2.0.1
parsel==1.1.0
Pillow==3.4.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
PyJWT==1.4.2
PyMySQL==0.7.10
pyOpenSSL==16.2.0
python-social-auth==0.3.3
python3-openid==3.0.10
pytz==2017.2
queuelib==1.4.2
readability-lxml==0.6.2
reportlab==3.3.0
requests==2.12.1
requests-oauthlib==0.7.0
Scrapy==1.3.0
service-identity==16.0.0
six==1.10.0
social-auth-app-django==0.1.0
social-auth-core==0.2.1
Twisted==16.6.0
w3lib==1.16.0
zope.interface==4.3.3
